requirements.txt
kivy==2.1.0
requests==2.31.0
Pillow==10.0.0